made a component edit button changes dynamically
